# openapi.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Malformed or invalid request"
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The referenced gallery or image was not found"
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "No verified Cloudflare Access session"
      },
      "Unavailable": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "A backing store is temporarily unavailable"
      },
      "Unprocessable": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The request was understood but the operation failed (bad source link, slug collision, ...)"
      }
    },
    "schemas": {
      "CreateRequest": {
        "additionalProperties": false,
        "properties": {
          "order": {
            "description": "Optional image refs (filenames for Dropbox) in the desired display order; unlisted images keep scan order after them",
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "type": "array"
          },
          "url": {
            "description": "A public, download-enabled Dropbox folder, Google Drive folder, or iCloud shared album URL to create the gallery from",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "type": "object"
      },
      "DeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "description": "Human-readable failure reason",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "GalleryList": {
        "additionalProperties": false,
        "properties": {
          "galleries": {
            "items": {
              "$ref": "#/components/schemas/Gallery"
            },
            "type": "array"
          }
        },
        "required": [
          "galleries"
        ],
        "type": "object"
      },
      "GalleryResponse": {
        "additionalProperties": false,
        "properties": {
          "gallery": {
            "$ref": "#/components/schemas/Gallery"
          }
        },
        "required": [
          "gallery"
        ],
        "type": "object"
      },
...
      "ScanRequest": {
        "additionalProperties": false,
        "properties": {
          "url": {
            "description": "A public, download-enabled Dropbox folder, Google Drive folder, or iCloud shared album URL",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "type": "object"
      },
      "ScanResponse": {
        "additionalProperties": false,
        "properties": {
          "scan": {
            "$ref": "#/components/schemas/GalleryScan"
          }
        },
        "required": [
          "scan"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cloudflareAccess": {
        "description": "A verified Cloudflare Access JWT, sent as the Cf-Access-Jwt-Assertion header or CF_Authorization cookie. The server verifies issuer, audience, signature, expiry, and the immutable sub claim before authorizing management operations.",
        "in": "header",
        "name": "Cf-Access-Jwt-Assertion",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Gallery management and source image APIs for Manorama, the photography-first publishing application. Management operations require a verified Cloudflare Access session; the image proxies are public because public gallery pages load images through them.",
    "title": "Manorama Gallery Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/dropbox/file": {
      "get": {
        "description": "Fetch a full-resolution image from a Dropbox folder by source URL and filename.",
        "operationId": "get_dropbox_file",
        "parameters": [
          {
            "description": "The Dropbox folder source URL",
            "in": "query",
            "name": "sourceUrl",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The image filename",
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The original image bytes, never recompressed"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [],
        "summary": "Fetch a Dropbox original",
        "tags": [
          "dropbox"
        ]
      }
    },
    "/api/dropbox/thumbnail": {
      "get": {
        "description": "Fetch a thumbnail image from a Dropbox folder by source URL and filename.",
        "operationId": "get_dropbox_thumbnail",
        "parameters": [
          {
            "description": "The Dropbox folder source URL",
            "in": "query",
            "name": "sourceUrl",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The image filename",
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The thumbnail image bytes"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [],
        "summary": "Fetch a Dropbox thumbnail",
        "tags": [
          "dropbox"
        ]
      }
    },
    "/api/drive/file": {
      "get": {
        "description": "Fetch a full-resolution image from a link-shared Google Drive folder by file ID.",
        "operationId": "get_drive_file",
        "parameters": [
          {
            "description": "The Google Drive file ID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The file's resource key, required for resource-keyed shared links",
            "in": "query",
            "name": "rk",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The original image bytes, never recompressed"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [],
        "summary": "Fetch a Google Drive original",
        "tags": [
          "gdrive"
        ]
      }
    },
    "/api/drive/thumbnail": {
      "get": {
        "description": "Fetch a thumbnail image from a link-shared Google Drive folder by file ID. HEIC files are served as JPEG renditions.",
        "operationId": "get_drive_thumbnail",
        "parameters": [
          {
            "description": "The Google Drive file ID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Thumbnail width class (w256 default, w2048 for HEIC previews)",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "enum": [
                "w256",
                "w2048"
              ],
              "type": "string"
            }
          },
          {
            "description": "The file's resource key, required for resource-keyed shared links",
            "in": "query",
            "name": "rk",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The thumbnail image bytes"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [],
        "summary": "Fetch a Google Drive thumbnail",
        "tags": [
          "gdrive"
        ]
      }
    },
    "/api/icloud/image": {
      "get": {
        "description": "Fetch an image derivative from a public iCloud shared album. The expiring CDN URL is re-resolved on every request.",
        "operationId": "get_icloud_image",
        "parameters": [
          {
            "description": "The shared album token",
            "in": "query",
            "name": "album",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The photo GUID within the album",
            "in": "query",
            "name": "photo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The derivative checksum selecting the rendition",
            "in": "query",
            "name": "c",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The derivative image bytes (JPEG, web resolution)"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [],
        "summary": "Fetch an iCloud shared album image",
        "tags": [
          "icloud"
        ]
      }
    },
    "/api/galleries": {
      "get": {
        "description": "List all photo galleries with titles, captions, dates, and image counts.",
        "operationId": "list_galleries",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GalleryList"
                }
              }
            },
            "description": "The gallery list, newest first"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "summary": "List galleries",
        "tags": [
          "galleries"
        ]
      },
      "post": {
        "description": "Scan a public Dropbox, Google Drive, or iCloud shared album link and create a gallery from its images, optionally in a given order.",
        "operationId": "create_gallery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GalleryResponse"
                }
              }
            },
            "description": "The created gallery"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "summary": "Create a gallery from a shared link",
        "tags": [
          "galleries"
        ]
      }
    },
    "/api/galleries/scan": {
      "post": {
        "description": "Scan a public Dropbox, Google Drive, or iCloud shared album link to discover images for a new gallery. Returns the source title and image metadata.",
        "operationId": "scan_gallery_source",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "The scan result"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "summary": "Scan a shared link",
        "tags": [
          "galleries"
        ]
      }
    },
    "/api/galleries/{slug}": {
      "delete": {
        "description": "Delete a gallery and its Manorama references. This is irreversible; source files are never touched.",
        "operationId": "delete_gallery",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Deletion confirmation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "summary": "Delete a gallery",
        "tags": [
          "galleries"
        ]
      },
      "parameters": [
        {
          "description": "Current gallery URL slug — the resource identity",
          "in": "path",
          "name": "slug",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "description": "Update a gallery's title, caption, image order, or URL slug (rename via newSlug).",
        "operationId": "update_gallery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GalleryResponse"
                }
              }
            },
            "description": "The updated gallery (under its post-rename slug when newSlug was applied)"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "summary": "Update a gallery",
        "tags": [
          "galleries"
        ]
      }
    },
    "/api/galleries/{slug}/refresh": {
      "parameters": [
        {
          "description": "Gallery URL slug to re-scan",
          "in": "path",
          "name": "slug",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "description": "Re-scan a gallery's source link and update its images with any new or changed files.",
        "operationId": "refresh_gallery",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GalleryResponse"
                }
              }
            },
            "description": "The refreshed gallery"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "summary": "Refresh a gallery from its source",
        "tags": [
          "galleries"
        ]
      }
    }
  },
  "security": [
    {
      "cloudflareAccess": []
    }
  ],
  "servers": [
    {
      "url": "https://manorama.xyz"
    }
  ],
...
}